third_party/blink/web_tests/fast/dom/shadow/event-path-window-load.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
setup({ explicit_done: true });

var img = document.getElementById("img");
window.onload = function (e) {
    var path = e.composedPath();
    test(function () {
        assert_equals(path.length, 1, 'path.length');
        assert_equals(path[0], window, 'path[0] should be window');
    }, "event.composedPath() for window.onload");
    done();
};
</script>
